live-preview: Extract the property_view related code from ui.rs
The file is just too big!

This patch does not contain any intended code changes.
